[WIP] fix test
